# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: ANAS Web API
  version: 0.8.0-m5
  description: |
    M2 API for the ANAS administration daemon. This contract is versioned
    independently from the CLI JSON contract: HTTP responses use
    `anas.dev/api/v1`, while the CLI continues to use `anas.dev/cli/v1`.

    Workspaces are registered when `anasd` starts. Callers address them only by
    opaque workspace ID; request and response DTOs never contain host filesystem
    paths. M1A adds bootstrap exchange and local-owner authentication while the
    workspace surface remains read-only. Route visibility and transport are
    state-dependent. M1B adds durable job history and event replay, the
    certificate-bound bootstrap-to-enrollment handoff and initial local-owner
    creation, synchronous managed-configuration read/validation/atomic save,
    audited typed deployment planning, bootstrap first-apply jobs, and local
    owner step-up for full-state direct apply. M1C adds a separately built Vue
    console shell plus an independent recovery bundle, both served from
    immutable assets embedded in `anasd`. M1.5 adds a
    separate mTLS trusted-proxy listener whose verified OIDC directory identity
    receives the same management capabilities while local-account routes remain
    direct-listener only. M2 adds live container health, dependency-expanded
    Module lifecycle previews and durable start/stop/restart/rollback jobs.
...
    Saving desired configuration does not apply it. A disallowed
...
servers:
...
security: []
...
paths:
...
  /api/v1/workspaces/{ws}/deployments:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
    get:
      operationId: listDeployments
      summary: List deployment state in reverse chronological order
      tags: [Deployments]
      security:
        - {}
        - localSessionCookie: []
      x-anas-permission: workspace.read
      x-anas-scope: workspace
      x-anas-listeners: [direct, trusted_proxy]
      x-anas-state-access:
        m0: {authentication: none, transports: [plaintext, tls]}
        full: {authentication: owner, transports: [tls]}
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: One page of deployment state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentListResponse"
        "400":
          $ref: "#/components/responses/BadRequestProblem"
        "401":
          $ref: "#/components/responses/UnauthorizedProblem"
        "403":
          $ref: "#/components/responses/ForbiddenProblem"
        "404":
          $ref: "#/components/responses/NotFoundProblem"
        "405":
          $ref: "#/components/responses/GetMethodNotAllowedProblem"
        "408":
          $ref: "#/components/responses/RequestCanceledProblem"
        "500":
          $ref: "#/components/responses/InternalProblem"
        "504":
          $ref: "#/components/responses/DeadlineProblem"
  /api/v1/workspaces/{ws}/deployments/{id}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
      - $ref: "#/components/parameters/DeploymentID"
    get:
      operationId: inspectDeployment
      summary: Inspect a deployment manifest and its state
      description: |
        Returns a deliberately safe projection of the deployment artifact. Host
        paths, hook commands, resource specifications, and secret-store keys are
        omitted even when they exist in the on-disk manifest. All digests derived
        from configuration, rendered trees, or other secret-bearing inputs are
        omitted as well, so the read API cannot act as an offline guessing oracle.
      tags: [Deployments]
      security:
        - {}
        - localSessionCookie: []
      x-anas-permission: workspace.read
      x-anas-scope: workspace
      x-anas-listeners: [direct, trusted_proxy]
      x-anas-state-access:
        m0: {authentication: none, transports: [plaintext, tls]}
        full: {authentication: owner, transports: [tls]}
      responses:
        "200":
          description: Safe deployment metadata and state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequestProblem"
        "401":
          $ref: "#/components/responses/UnauthorizedProblem"
        "403":
          $ref: "#/components/responses/ForbiddenProblem"
        "404":
          $ref: "#/components/responses/NotFoundProblem"
        "405":
          $ref: "#/components/responses/GetMethodNotAllowedProblem"
        "408":
          $ref: "#/components/responses/RequestCanceledProblem"
        "412":
          $ref: "#/components/responses/PreconditionProblem"
        "500":
          $ref: "#/components/responses/InternalProblem"
        "504":
          $ref: "#/components/responses/DeadlineProblem"
  /api/v1/workspaces/{ws}/modules/{module}/commands:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
      - $ref: "#/components/parameters/ModuleName"
    get:
      operationId: listModuleCommands
      summary: List commands frozen for one Module in the active deployment
      description: Returns only public descriptors and local availability; it does not contact the managed service.
      tags: [Modules]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      security:
        - {}
        - localSessionCookie: []
      x-anas-permission: workspace.read
      x-anas-scope: workspace
      x-anas-listeners: [direct, trusted_proxy]
      x-anas-state-access:
        m0: {authentication: none, transports: [plaintext, tls]}
        full: {authentication: owner, transports: [tls]}
      responses:
        "200":
          description: Public Module Command descriptors in stable command-ID order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleCommandListResponse"
        "400":
          $ref: "#/components/responses/BadRequestProblem"
        "401":
          $ref: "#/components/responses/UnauthorizedProblem"
        "403":
          $ref: "#/components/responses/ForbiddenProblem"
        "404":
          $ref: "#/components/responses/NotFoundProblem"
        "405":
          $ref: "#/components/responses/GetMethodNotAllowedProblem"
        "408":
          $ref: "#/components/responses/RequestCanceledProblem"
        "412":
          $ref: "#/components/responses/PreconditionProblem"
        "500":
          $ref: "#/components/responses/InternalProblem"
        "504":
          $ref: "#/components/responses/DeadlineProblem"
  /api/v1/workspaces/{ws}/modules/{module}/commands/{command}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
      - $ref: "#/components/parameters/ModuleName"
      - $ref: "#/components/parameters/ModuleCommandID"
    get:
      operationId: getModuleCommand
      summary: Read one command frozen for a Module in the active deployment
      tags: [Modules]
      security:
        - {}
        - localSessionCookie: []
      x-anas-permission: workspace.read
      x-anas-scope: workspace
      x-anas-listeners: [direct, trusted_proxy]
      x-anas-state-access:
        m0: {authentication: none, transports: [plaintext, tls]}
        full: {authentication: owner, transports: [tls]}
      responses:
        "200":
          description: One public Module Command descriptor and its local availability.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleCommandDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequestProblem"
        "401":
          $ref: "#/components/responses/UnauthorizedProblem"
        "403":
          $ref: "#/components/responses/ForbiddenProblem"
        "404":
          $ref: "#/components/responses/NotFoundProblem"
        "405":
          $ref: "#/components/responses/GetMethodNotAllowedProblem"
        "408":
          $ref: "#/components/responses/RequestCanceledProblem"
        "412":
          $ref: "#/components/responses/PreconditionProblem"
        "500":
          $ref: "#/components/responses/InternalProblem"
        "504":
          $ref: "#/components/responses/DeadlineProblem"
...
components:
  securitySchemes:
    preAuthCSRFCookie:
      type: apiKey
      in: cookie
      name: anas_preauth_csrf
      description: |
        Host-only double-submit cookie issued by `GET /api/v1/auth/csrf`.
        Its value must exactly match `X-CSRF-Token`.
      x-anas-cookie-attributes:
        host_only: true
        http_only: true
        same_site: Strict
        path: /
        secure: state-dependent
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: The matching pre-authentication/session CSRF value; enrollment reads it from the dedicated same-origin CSRF cookie rather than a JSON exchange response.
    bootstrapSessionCookie:
      type: apiKey
      in: cookie
      name: anas_bootstrap_session
      description: |
        Host-only bootstrap session credential issued by token exchange. It is
        HttpOnly and SameSite=Strict, has Path=/ and no Domain, and is Secure
        when issued in enrollment. A promoted session authorizes only the
        enrollment handoff route for its original transaction.
      x-anas-cookie-attributes:
        host_only: true
        http_only: true
        same_site: Strict
        path: /
        secure: state-dependent
    enrollmentSessionCookie:
      type: apiKey
      in: cookie
      name: __Host-anas_enrollment_session
      description: |
        Host-only enrollment credential issued only at the validated TLS target.
        It is Secure, HttpOnly and SameSite=Strict, has Path=/ and no Domain.
...
  parameters:
...
    WorkspaceID:
      name: ws
      in: path
      required: true
      description: Opaque ID of a workspace registered when the daemon started; never a filesystem path.
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$"
    DeploymentID:
      name: id
      in: path
      required: true
      description: Deployment artifact ID, not a path.
      schema:
        type: string
        minLength: 1
        maxLength: 255
    ModuleName:
      name: module
      in: path
      required: true
      description: Module name in the active deployment.
      schema:
        type: string
        pattern: "^[a-z][a-z0-9_]*$"
...
    ModuleCommandID:
      name: command
      in: path
      required: true
      description: Stable Module-local command ID.
      schema:
        type: string
        minLength: 1
        maxLength: 63
        pattern: "^[a-z][a-z0-9-]{0,62}$"
...
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of list items to return.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor returned as `next_cursor` by the preceding page.
      schema:
        type: string
        minLength: 1
        maxLength: 512
...
  responses:
...
    BadRequestProblem:
      description: The Host or Origin is invalid, a path or body value is invalid, or the query string is malformed or unsupported.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    UnauthorizedProblem:
      description: The credential is missing, expired, invalid, or otherwise unauthenticated.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ForbiddenProblem:
      description: Origin, CSRF, or authorization validation failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFoundProblem:
      description: The resource does not exist, or the route is hidden for the current state, listener, or transport.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    PreconditionProblem:
      description: A supplied entity validator is not strong and current, or persisted state fails its required integrity precondition.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    GetMethodNotAllowedProblem:
      description: The route only accepts GET in the current state and transport.
      headers:
        Allow:
          required: true
          schema:
            type: string
            const: GET
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    RequestCanceledProblem:
      description: Request processing was canceled.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    InternalProblem:
      description: The daemon could not read state, authorize, map, or encode the response.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    DeadlineProblem:
      description: The request deadline expired before the query completed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    APIVersion:
      type: string
      const: anas.dev/api/v1
...
    DeploymentListResponse:
      type: object
      additionalProperties: false
      required: [api_version, workspace_id, items, next_cursor]
      properties:
        api_version:
          $ref: "#/components/schemas/APIVersion"
        workspace_id:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/DeploymentState"
        next_cursor:
          $ref: "#/components/schemas/NullableString"
    DeploymentDetailResponse:
      type: object
      additionalProperties: false
      required: [api_version, workspace_id, deployment, state]
      properties:
        api_version:
          $ref: "#/components/schemas/APIVersion"
        workspace_id:
          type: string
        deployment:
          $ref: "#/components/schemas/Deployment"
        state:
          $ref: "#/components/schemas/DeploymentState"
...
    ModuleCommandListResponse:
      type: object
      additionalProperties: false
      required: [api_version, workspace_id, active_deployment, items, next_cursor]
      properties:
        api_version:
          $ref: "#/components/schemas/APIVersion"
        workspace_id:
          type: string
        active_deployment:
          $ref: "#/components/schemas/NullableString"
        items:
          type: array
          items:
            $ref: "#/components/schemas/ModuleCommand"
        next_cursor:
          $ref: "#/components/schemas/NextCursor"
...